Anne Updates
created MaukaRainGauge (& renamed RainGauge to MakaiRainGauge). still not imported in rain-conditions.tsx because i'm waiting for Brian's code so i can make a new backend and put everything together.
changed the min and max levels of the gauge charts and line graphs according to the spreadsheet. added tick labels to all the gauge charts based on their respective thresholds.